fix: bug in generating orderline record key

+++ app/src/initializer.cpp
@@ -70,7 +70,7 @@ void create_and_insert_orderline_record(
 };
 
 void load_items_table() {
-    for (uint32_t i_id = 1; i_id <= Item::ITEMS; i_id++) {
+    for (int i_id = 1; i_id <= Item::ITEMS; i_id++) {
       create_and_insert_item_record(i_id);
     }
 }
@@ -78,7 +78,7 @@ void load_items_table() {
 // Loading warehouses table eventually evokes loading of all the tables other than the items table.
 void load_warehouses_table() {
   const Config& c = get_config();
-    for (uint16_t w_id = 1; w_id <= c.get_num_warehouses(); w_id++) {
+    for (int w_id = 1; w_id <= c.get_num_warehouses(); w_id++) {
         create_and_insert_warehouse_record(w_id);
         load_stocks_table(w_id);
         load_districts_table(w_id);
@@ -86,13 +86,13 @@ void load_warehouses_table() {
 }
 
 void load_stocks_table(uint16_t s_w_id) {
-    for (uint32_t s_id = 1; s_id <= Stock::STOCKS_PER_WARE; s_id++) {
+    for (int s_id = 1; s_id <= Stock::STOCKS_PER_WARE; s_id++) {
         create_and_insert_stock_record(s_w_id, s_id);
     }
 }
 
 void load_districts_table(uint16_t d_w_id) {
-    for (uint8_t d_id = 1; d_id <= District::DISTS_PER_WARE; d_id++) {
+    for (int d_id = 1; d_id <= District::DISTS_PER_WARE; d_id++) {
         create_and_insert_district_record(d_w_id, d_id);
         load_customers_table(d_w_id, d_id);
         load_orders_table(d_w_id, d_id);
@@ -101,7 +101,7 @@ void load_districts_table(uint16_t d_w_id) {
 
 void load_customers_table(uint16_t c_w_id, uint8_t c_d_id) {
   Timestamp t = get_timestamp();
-    for (uint32_t c_id = 1; c_id <= Customer::CUSTS_PER_DIST; c_id++) {
+    for (int c_id = 1; c_id <= Customer::CUSTS_PER_DIST; c_id++) {
         create_and_insert_customer_record(c_w_id, c_d_id, c_id, t);
         load_histories_table(c_w_id, c_d_id, c_id);
     }
@@ -114,7 +114,7 @@ void load_histories_table(uint16_t w_id, uint8_t d_id, uint32_t c_id) {
 
 void load_orders_table(uint16_t o_w_id, uint8_t o_d_id) {
     Permutation p(1, Order::ORDS_PER_DIST);
-    for (uint32_t o_id = 1; o_id <= Order::ORDS_PER_DIST; o_id++) {
+    for (int o_id = 1; o_id <= Order::ORDS_PER_DIST; o_id++) {
         uint32_t o_c_id = p[o_id - 1];
         Timestamp o_entry_d = create_and_insert_order_record(o_w_id, o_d_id, o_c_id, o_id);
 
@@ -127,10 +127,10 @@ void load_orders_table(uint16_t o_w_id, uint8_t o_d_id) {
 
 void load_orderlines_table(uint16_t ol_w_id, uint8_t ol_d_id, uint32_t ol_o_id, Timestamp o_entry_d) {
     uint8_t ol_cnt = urand_int(OrderLine::MIN_ORDLINES_PER_ORD, OrderLine::MAX_ORDLINES_PER_ORD);
-    for (uint8_t ol_number = 1; ol_number <= ol_cnt; ol_number++) {
+    for (int ol_number = 1; ol_number <= ol_cnt; ol_number++) {
         uint32_t ol_i_id = urand_int(1, 100000);
         create_and_insert_orderline_record(
-            ol_w_id, ol_d_id, ol_o_id, ol_w_id, ol_i_id, ol_number, o_entry_d);
+                                           ol_w_id, ol_d_id, ol_o_id, ol_w_id, ol_i_id, ol_number, o_entry_d);
     }
 }
